move hhuc to hhu
