+++ docs/university.md
@@ -591,11 +591,9 @@ category 列表：
 
 <Route author="plusmultiply0" example="/hhu/libNews" path="/hhu/libNews"/>
 
-## 河海大学（常州校区）
-
 ### 河海大学常州校区图书馆 - 新闻动态
 
-<Route author="plusmultiply0" example="/hhuc/libNews" path="/hhuc/libNews"/>
+<Route author="plusmultiply0" example="/hhu/libNewsc" path="/hhuc/libNewsc"/>
 
 ## 河南大学
