+++ lib/router.js
@@ -585,7 +585,7 @@ router.get('/nuaa/yjsy/:type?', require('./routes/universities/nuaa/yjsy/yjsy'))
 // 河海大学
 router.get('/hhu/libNews', require('./routes/universities/hhu/libNews'));
 // 河海大学常州校区
-router.get('/hhuc/libNews', require('./routes/universities/hhuc/libNews'));
+router.get('/hhu/libNewsc', require('./routes/universities/hhu/libNewsc'));
 
 // 哈尔滨工业大学
 router.get('/hit/jwc', require('./routes/universities/hit/jwc'));
